content/canvas/test/test_2d.shadow.attributes.shadowBlur.1.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.shadow.attributes.shadowBlur.1</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.shadowBlur = 1;
ok(ctx.shadowBlur === 1, "ctx.shadowBlur === 1");
ctx.shadowBlur = 0.5;
ok(ctx.shadowBlur === 0.5, "ctx.shadowBlur === 0.5");
ctx.shadowBlur = 1e6;
ok(ctx.shadowBlur === 1e6, "ctx.shadowBlur === 1e6");
ctx.shadowBlur = 1;
ctx.shadowBlur = -2;
ok(ctx.shadowBlur === 1, "ctx.shadowBlur === 1");
ctx.shadowBlur = 0;
ok(ctx.shadowBlur === 0, "ctx.shadowBlur === 0");

SimpleTest.finish();

});
</script>
